+++ app/src/main/java/org/simple/clinic/newentry/form/ZoneField.kt
@@ -2,7 +2,9 @@ package org.simple.clinic.newentry.form
 
 import org.simple.clinic.newentry.form.ValidationError.MissingValue
 
-class ZoneField(labelResId: Int) : InputField<String>(labelResId) {
+data class ZoneField(
+    private val _labelResId: Int
+) : InputField<String>(_labelResId) {
   override fun validate(value: String): Set<ValidationError> {
     return if (value.isBlank()) setOf(MissingValue) else emptySet()
   }