+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 ### Internal
 - Migrated `LoggedOutOfDeviceDialog` to Mobius
+- Setup the UI via an effect in `PatientEntryScreen`
 - [In Progress: 06 Aug 2020] Migrate `PatientSearchResultsScreen` to Mobius
 - [In Progress: 10 Aug 2020] Migrate `ConfirmRemoveBloodPressureDialog` to Mobius
 - [In Progress: 11 Aug 2020] Migrate `TheActivity` to Mobius